Automatically chooses cenovnik based on it's date
You don't have to choose cenovnik if you want to add products
to faktura. Backend does that automatically basically:
if there are no valid cenovnik - nothing is returned
if there are cenovniks but past it's due - nothing
if there is valid cenovnik - roba in it is returned and can
be chosen when using faktura
if there are 2 valid cenovniks - app will choose the one that
will expire sooner.